requirements.txt
pytest==5.2.1
pytest_bdd
selenium
jsonpickle
pony
pymysql
#mysql-connector-python-rf
lxml-4.4.1-cp37-cp37m-win_amd64.whl
allure-pytest==2.8.6
